Add ARIA attributes for accessability in jquery code.
ToDo: update prototype code to match.
Related to https://github.com/harvesthq/chosen/issues/264
Use the label element, aria-label, or aria-labelledby attributes from the select box for the text input field.
Add attributes and ids such that a screen reader can know which element in the search results is highlighted so that it can be read:
    Attributes on the text input:
      aria-haspopup="true"
      role="combobox"
      aria-autocomplete="list"
      aria-owns="id_of_list_of_options"
      aria-activedescendant="id_of_highlighted_option" (change this value programatically when the highlighted option changes)
    Attributes on the list of options:
      id for use in the aria-owns attribute on the input
    Attributes on each option:
      role="option"
      id for use in the aria-activedescendant attribute on the input